requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.5.2; python_version >= '3.7'
asttokens==2.0.5
backcall==0.2.0
certifi==2022.5.18.1; python_version >= '3.6'
cffi==1.15.0
charset-normalizer==2.0.12; python_full_version >= '3.5.0'
cryptography==37.0.2; python_version >= '3.6'
decorator==5.1.1; python_full_version >= '3.5.0'
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
dj-database-url==0.5.0
dj-rest-auth==2.2.4
django-allauth==0.51.0
django-auth==0.1.9
django-cors-headers==3.13.0
django-extensions==3.1.5
django-heroku==0.3.1
django-templated-mail==1.1.1
django==3.2.10
djangorestframework-simplejwt==5.2.0
djangorestframework==3.13.1
djoser==2.0.5
drf-writable-nested==0.6.3
executing==0.8.3
gunicorn==20.1.0
idna==3.3; python_full_version >= '3.5.0'
ipython==8.4.0
jedi==0.18.1; python_version >= '3.6'
matplotlib-inline==0.1.3; python_full_version >= '3.5.0'
oauthlib==3.2.0; python_version >= '3.6'
parso==0.8.3; python_version >= '3.6'
pexpect==4.8.0; sys_platform != 'win32'
pickleshare==0.7.5
prompt-toolkit==3.0.29; python_full_version >= '3.6.2'
psycopg2-binary==2.9.3
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pygments==2.12.0; python_version >= '3.6'
pyjwt[crypto]==2.4.0; python_version >= '3.6'
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2022.1
requests-oauthlib==1.3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
requests==2.28.0; python_version >= '3.7' and python_version < '4'
rest-social-auth==7.0.0
setuptools==62.4.0; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
social-auth-app-django==5.0.0
social-auth-core==4.3.0; python_version >= '3.6'
sqlparse==0.4.2; python_full_version >= '3.5.0'
stack-data==0.2.0
traitlets==5.2.2.post1; python_version >= '3.7'
urllib3==1.26.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
wcwidth==0.2.5
whitenoise==6.2.0; python_version >= '3.7'
